requirements.txt
# AWS SDK
boto3==1.34.19
botocore==1.34.19

# OpenAI
openai==1.7.2

# Email parsing
email-validator==2.1.0

# Document parsing
pdfplumber==0.10.3
PyMuPDF==1.23.8
python-docx==1.1.0
Pillow==10.1.0
openpyxl==3.1.2

# OCR (optional - requires Tesseract installed)
pytesseract==0.3.10

# Utilities
python-dotenv==1.0.0
requests==2.31.0

